Day 3. Removed the timing bobbins as it was crufty
First pass was using 1..n which worked for part1 but failed on part2. So
changed it to b 0 based and it worked. Can't be arsed to work out what
was wrong.

Couldn't work out a nice way of initialising a set with the pairs for
each slope so just hard coded the calls.