Day 8, both parts. Simple CPU thang.
Tweaked things in the progression from part1 to part2. Originally the
code was passed in to the cpu constructor. But as for part2 we want to
re-run slightly different code it makes more sense to construct one CPU
and pass the code in to the execute. Also separated out Parse so it
could be used by Part2 directly.